+++ includes/classes/Setup.php
@@ -39,6 +39,7 @@ class Setup
   require_once 'ewp-gallery/class-ewp-gallery.php';
   require_once 'ewp-fields/class-field.php';
   require_once 'ewp-wp-content/class-wp-content.php';
+  require_once 'ewp-wp-content/class-slug-manager.php';
   require_once 'ewp-wp-content/class-wp-content-installer.php';
   require_once 'ewp-search-filter/class-wp-search.php';
   require_once 'awm-api/class-awm-api.php';